app/requirements.txt
numpy==1.23.5
streamlit==1.22.0
Pillow==9.5.0
keras==2.12.0
tensorflow==0.12.0